requirements.txt
alembic==1.17.2
aniso8601==10.0.1
attrs==25.4.0
blinker==1.9.0
click==8.3.1
Flask==3.1.2
Flask-Admin==2.0.2
Flask-DebugToolbar==0.16.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
flask-restx==1.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
gunicorn==23.0.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
pillow==12.0.0
python-dotenv==1.2.1
referencing==0.37.0
rpds-py==0.30.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.4
WTForms==3.2.1
